requirements.txt
nltk==3.4.1
PyYAML==5.1
six==1.12.0
